Add output to check_if_version_is_incremented
Make this workflow more flexible by outputting the new package version
once it completes. This allows the workflow to be used eg to draft or
publish new releases, create git tags etc.

Another logic change is to support pushes on main branch. When run on
the main branch, the workflow now uses the previous commit on this
branch as a base to compare new version vs old.